Update font.h
Changed the type of "u8 color" to "u16 color" in font_char and font_str.

+++ kernel/ui/system/font.h
@@ -21,8 +21,8 @@ extern "C" {
         font_str(__s, __x, __y, __c);\
     } while (0);
 
-void font_char(char c, size_t x, size_t y, u8 color, u8 scale);
-void font_str(const char *s, size_t x, size_t y, u8 color, u8 scale);
+void font_char(char c, size_t x, size_t y, u16 color, u8 scale);
+void font_str(const char *s, size_t x, size_t y, u16 color, u8 scale);
 
 #ifdef __cplusplus
 }
